pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "kettle"
version = "0.1.0"
description = "Build-time verification and attestation for TEE (Trusted Execution Environment) deployments"
requires-python = ">=3.10"

dependencies = [
    "requests>=2.32.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "pymerkle>=6.0.0",
    "numpy>=1.24.0",
    "safetensors>=0.4.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
service = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.12",
]

[project.scripts]
kettle = "kettle.cli:main"
kettle-service = "service.app:app"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
